add profile page and move calendar page to index
