+++ pages/index.tsx
@@ -1,58 +1,35 @@
 import Link from "next/link";
-import React, { useEffect, useContext, useState } from "react";
+import React from "react";
+import { useContext, useEffect, useState } from "react";
 import Layout from "../components/Layout";
+import { UserCalendar } from "../components/UserCalendar";
 import { AuthContext } from "../context/AuthContext";
-import { GroupWithId, loadGroup } from "../interfaces/Group";
-import { loadUser, User } from "../interfaces/User";
+import {
+  DateStatusList,
+  loadDateStatusList,
+  storeDateStatusList,
+} from "../interfaces/DateStatus";
 
 const IndexPage = (): JSX.Element => {
   const { authUser } = useContext(AuthContext);
-  const [user, setUser] = useState<User | null | undefined>(undefined);
-  const [groups, setGroups] = useState<GroupWithId[] | null | undefined>(
-    undefined
-  );
-
-  // TODO よく使う処理なのでカスタムフックにする
+  const [dateStatusList, setDateStatusList] = useState<
+    DateStatusList | undefined | null
+  >(undefined);
   useEffect(() => {
     // コンポーネントが削除された後にsetDateStatusListが呼ばれないようにするため
     let unmounted = false;
     const setFromDatabase = async () => {
       if (authUser == null) {
         if (!unmounted) {
-          setUser(null);
-          setGroups(null);
+          setDateStatusList(null);
         }
       } else {
         // TODO エラー処理
-        const user = await loadUser(authUser.uid);
-        if (user != null && !unmounted) {
-          setUser(user);
-          const groupList: GroupWithId[] = [];
-          if (user.groups != null) {
-            const groupIds = Object.keys(user.groups);
-            const groupsfromDatabase = await Promise.all(
-              groupIds.map((groupId) => loadGroup(groupId))
-            );
-            for (const group of groupsfromDatabase) {
-              if (group != null) {
-                groupList.push(group);
-              }
-            }
-          }
-          if (!unmounted) {
-            groupList.sort((a, b) => {
-              const nameA = a.name.toUpperCase();
-              const nameB = b.name.toUpperCase();
-              if (nameA < nameB) {
-                return -1;
-              }
-              if (nameA > nameB) {
-                return 1;
-              }
-              return 0;
-            });
-            setGroups(groupList);
-          }
+        const data = await loadDateStatusList(authUser.uid);
+        if (data != null && !unmounted) {
+          setDateStatusList(data);
+        } else if (data == null && !unmounted) {
+          setDateStatusList({});
         }
       }
     };
@@ -64,96 +41,47 @@ const IndexPage = (): JSX.Element => {
     };
     return cleanup;
   }, [authUser]);
-
-  const groupListComponent = (groupList: GroupWithId[]): JSX.Element => {
-    const component = groupList.map((g) => {
-      if (g.invitationId == null) {
-        return (
-          <div key={g.id}>
-            <div>
-              <Link href="/groups/[id]" as={`/groups/${g.id}`}>
-                <a>{g.name}</a>
-              </Link>
-            </div>
-            <div>
-              <Link
-                href="/create-invitation/[id]"
-                as={`/create-invitation/${g.id}`}
-              >
-                <a>招待リンク作成</a>
-              </Link>
-            </div>
-          </div>
-        );
-      } else {
-        return (
-          <div key={g.id}>
-            <div>
-              <Link href="/groups/[id]" as={`/groups/${g.id}`}>
-                <a>{g.name}</a>
-              </Link>
-            </div>
-            <div>
-              <Link
-                href="/invitations/[id]"
-                as={`/invitations/${g.invitationId}`}
-              >
-                <a>招待リンク確認</a>
-              </Link>
-            </div>
-          </div>
-        );
+  useEffect(() => {
+    const setToDatabase = async () => {
+      if (authUser != null && dateStatusList != null) {
+        // TODO エラー処理
+        // TODO 失敗したらカレンダーの色は変えない
+        await storeDateStatusList(dateStatusList, authUser.uid);
       }
-    });
-    return (
-      <React.Fragment>
-        <p>groups:</p>
-        {component}
-      </React.Fragment>
-    );
-  };
-
-  // デバッグ用にユーザー情報を表示
-  // TODO ユーザーページにはSSRを使用
+    };
+    if (authUser !== undefined) {
+      setToDatabase();
+    }
+  }, [dateStatusList]);
   return (
-    <Layout title="Hima Share">
-      <h1>Hello Hima Share 👋</h1>
-      {user && groups && (
-        <React.Fragment>
-          <p>User info</p>
-          <p>name: {user.name}</p>
-          <p>email: {user.email}</p>
-          {groupListComponent(groups)}
-          <p>
-            <Link href="/calendar">
-              <a>カレンダー</a>
-            </Link>
-          </p>
-          <p>
-            <Link href="/create-group">
-              <a>グループ作成</a>
-            </Link>
-          </p>
-        </React.Fragment>
-      )}
-      {user === null && (
-        <React.Fragment>
-          <p className="text-main font-weight-bold">
-            You are not logged in yet
-          </p>
+    <React.Fragment>
+      {dateStatusList === null && (
+        <Layout title="Hima Share">
+          <h1>Hello Hima Share 👋</h1>
+          <h2>トップページ制作中...</h2>
           <p>
             <Link href="/login">
-              <a>Login</a>
+              <a>ログイン</a>
             </Link>
           </p>
           <p>
             <Link href="/register">
-              <a>Register</a>
+              <a>登録</a>
             </Link>
           </p>
-        </React.Fragment>
+        </Layout>
       )}
-    </Layout>
+      {dateStatusList && (
+        <Layout title="ユーザーカレンダー">
+          <h1>ユーザーカレンダー</h1>
+          <p>あなたの予定</p>
+          <UserCalendar
+            dateStatusList={dateStatusList}
+            setDateStatusList={(list) => setDateStatusList(list)}
+          />
+        </Layout>
+      )}
+    </React.Fragment>
   );
 };
